feat: add suggest to institutions

+++ src/components/Institution/institution.actions.js
@@ -3,18 +3,19 @@ import { FormattedMessage, injectIntl } from 'react-intl';
 import { Dropdown, Menu, Modal, Icon, Input, Button, Checkbox } from 'antd';
 import PropTypes from 'prop-types';
 import injectSheet from 'react-jss';
+import { Link } from 'react-router-dom';
 
 // API
-import { deleteInstitution, updateInstitution, mergeInstitutions, convertToCollection } from '../../api/institution';
+import { suggestConvertInstitution, suggestDeleteInstitution, suggestMergeInstitution, deleteInstitution, updateInstitution, mergeInstitutions, convertToCollection } from '../../api/institution';
 import { canDelete, canCreate, canUpdate } from '../../api/permissions';
 // Wrappers
 import withContext from '../hoc/withContext';
 // Components
 import { InstitutionSuggestWithoutContext as InstitutionSuggest } from '../common';
 
-const styles = {
+const { TextArea } = Input;
 
-};
+const styles = {};
 
 /**
  * Institution Actions component
@@ -30,7 +31,8 @@ class InstitutionActions extends React.Component {
     super(props);
 
     this.state = {
-      mergeWithInstitution: undefined
+      mergeWithInstitution: undefined,
+      proposerEmail: props.user ? props.user.email : null
     };
   }
 
@@ -59,7 +61,7 @@ class InstitutionActions extends React.Component {
     const hasConvertToCollection = await canCreate('grscicoll/institution', this.props.institution.key, 'convertToCollection');
     if (this._isMount) {
       // update state
-      this.setState({hasDelete, hasUpdate, hasMerge, hasConvertToCollection});
+      this.setState({ hasDelete, hasUpdate, hasMerge, hasConvertToCollection });
     };
     //else the component is unmounted and no updates should be made
   }
@@ -74,20 +76,34 @@ class InstitutionActions extends React.Component {
         </Menu.Item>
       )}
       {!institution.deleted && (
-        <Menu.Item key="delete" disabled={!this.state.hasDelete}>
+        <Menu.Item key="delete">
           <FormattedMessage id="delete.institution" defaultMessage="Delete this institution" />
+          {!this.state.hasDelete && <span style={{ color: '#aaa', marginLeft: 8 }}>
+            <FormattedMessage id="suggest" defaultMessage="Suggest" />
+          </span>}
         </Menu.Item>
       )}
-      {this.state.hasMerge && (
+      {!institution.deleted && (
         <Menu.Item key="merge">
           <FormattedMessage id="institution.merge" defaultMessage="Merge with other institution" />
+          {!this.state.hasMerge && <span style={{ color: '#aaa', marginLeft: 8 }}>
+            <FormattedMessage id="suggest" defaultMessage="Suggest" />
+          </span>}
         </Menu.Item>
       )}
-      {this.state.hasConvertToCollection && (
+      {!institution.deleted && (
         <Menu.Item key="convert">
           <FormattedMessage id="institution.convertToCollection" defaultMessage="Convert to collection" />
+          {!this.state.hasConvertToCollection && <span style={{ color: '#aaa', marginLeft: 8 }}>
+            <FormattedMessage id="suggest" defaultMessage="Suggest" />
+          </span>}
         </Menu.Item>
       )}
+      <Menu.Item key="create">
+        <Link to={`/institution/create`}>
+          <FormattedMessage id="institution.create" defaultMessage="Create new institution" />
+        </Link>
+      </Menu.Item>
     </Menu>;
   };
 
@@ -97,10 +113,20 @@ class InstitutionActions extends React.Component {
 
     switch (actionType) {
       case 'delete': {
-        title = intl.formatMessage({
-          id: 'delete.confirmation.institution',
-          defaultMessage: 'Are you sure to delete this institution?'
-        });
+        title = <div>
+          {intl.formatMessage({
+            id: 'delete.confirmation.collection',
+            defaultMessage: 'Are you sure you want to delete this collection? '
+          })
+          }
+          {!this.state.hasDelete && <div>
+            {intl.formatMessage({
+              id: "suggest.youCanOnlySuggest",
+              defaultMessage: "You can only suggest this action"
+            })
+            }
+          </div>}
+        </div>;
         break;
       }
       case 'restore': {
@@ -217,6 +243,24 @@ class InstitutionActions extends React.Component {
     });
   };
 
+  showSuggestConfirm = ({ title, action }) => {
+    // const { intl, user } = this.props;
+    // const description = intl.formatMessage({ id: 'collection.merge.comment', defaultMessage: 'This collection will be deleted after merging.' });
+    // const mergeLabel = intl.formatMessage({ id: 'merge', defaultMessage: 'Merge' });
+    // const cancelLabel = intl.formatMessage({ id: 'cancel', defaultMessage: 'Cancel' });
+    Modal.confirm({
+      title,
+      okText: 'Send suggestion',
+      okType: 'primary',
+      cancelText: 'Cancel',
+      content: <div>
+        <Input onChange={e => this.setState({ proposerEmail: e.target.value })} defaultvalue={this.state.proposerEmail} type="text" placeholder="email" style={{ marginBottom: 12 }}></Input>
+        <TextArea onChange={e => this.setState({ suggestComment: e.target.value })} type="text" placeholder="comment"></TextArea>
+      </div>,
+      onOk: action
+    });
+  };
+
   callAction = actionType => {
     switch (actionType) {
       case 'delete':
@@ -233,7 +277,20 @@ class InstitutionActions extends React.Component {
   merge = () => {
     const { institution, onChange } = this.props;
     const { mergeWithInstitution } = this.state;
-    mergeInstitutions({ institutionKey: institution.key, mergeIntoInstitutionKey: mergeWithInstitution }).then(() => onChange(null, 'crawl')).catch(onChange);
+    if (this.state.hasDelete) {
+      mergeInstitutions({ institutionKey: institution.key, mergeIntoInstitutionKey: mergeWithInstitution }).then(() => onChange(null, 'crawl')).catch(onChange);
+    } else {
+      this.showSuggestConfirm({
+        title: this.props.intl.formatMessage({id:"suggestion.pleaseProvideEmailAndComment", defaultMessage:'You are about to leave a suggestion, please provide your email and a comment'}),
+        action: () => {
+          suggestMergeInstitution({ mergeTargetKey: mergeWithInstitution, entityKey: institution.key, comments: [this.state.suggestComment], proposerEmail: this.state.proposerEmail })
+          .then(() => this.props.addSuccess({ statusText: this.props.intl.formatMessage({id:"suggestion.suggestionLogged", defaultMessage:"Thank you. Your suggestion has been logged"}) }))
+            .catch(error => {
+              this.props.addError({ status: error.response.status, statusText: error.response.data });
+            })
+        }
+      });
+    }
   };
 
   convert = () => {
@@ -250,7 +307,28 @@ class InstitutionActions extends React.Component {
     } else {
       body.institutionForNewCollectionKey = convertConfig.value;
     }
-    convertToCollection({ institutionKey: institution.key, body }).then(() => onChange(null, 'crawl')).catch(onChange);
+
+    if (this.state.hasConvertToCollection) {
+      convertToCollection({ institutionKey: institution.key, body }).then(() => onChange(null, 'crawl')).catch(onChange);
+    } else {
+      this.showSuggestConfirm({
+        title: this.props.intl.formatMessage({id: "suggestion.pleaseProvideEmailAndComment", defaultMessage: 'You are about to leave a suggestion, please provide your email and a comment'}),
+        action: () => {
+          suggestConvertInstitution({
+            institutionForConvertedCollection: body.institutionForNewCollectionKey,
+            nameForNewInstitutionForConvertedCollection: body.nameForNewInstitution,
+            entityKey: institution.key,
+            comments: [this.state.suggestComment],
+            proposerEmail: this.state.proposerEmail
+          })
+            .then(() => this.props.addSuccess({ statusText: this.props.intl.formatMessage({id:"suggestion.suggestionLogged", defaultMessage:"Thank you. Your suggestion has been logged"}) }))
+            .catch(error => {
+              this.props.addError({ status: error.response.status, statusText: error.response.data });
+            })
+        }
+      });
+    }
+
   };
 
   restoreItem = () => {
@@ -261,13 +339,28 @@ class InstitutionActions extends React.Component {
 
   deleteItem = () => {
     const { institution, onChange } = this.props;
-    deleteInstitution(institution.key).then(() => onChange()).catch(onChange);
+    if (this.state.hasDelete) {
+      deleteInstitution(institution.key).then(() => onChange()).catch(onChange);
+    } else {
+      this.showSuggestConfirm({
+        title: this.props.intl.formatMessage({id: "suggestion.pleaseProvideEmailAndComment", defaultMessage: 'You are about to leave a suggestion, please provide your email and a comment'}),
+        action: () => {
+          suggestDeleteInstitution({ entityKey: institution.key, comments: [this.state.suggestComment], proposerEmail: this.state.proposerEmail })
+          .then(() => this.props.addSuccess({ statusText: this.props.intl.formatMessage({id:"suggestion.suggestionLogged", defaultMessage:"Thank you. Your suggestion has been logged"}) }))
+            .catch(error => {
+              this.props.addError({ status: error.response.status, statusText: error.response.data });
+            })
+        }
+      });
+    }
   };
 
   render = () => {
     return (
       <Dropdown overlay={this.renderActionMenu()} arrow>
-        <Button><Icon type="more" /></Button>
+        <Button><Icon type="more" />
+          <FormattedMessage id="more" defaultMessage="More" />
+        </Button>
       </Dropdown>
     );
   }
@@ -279,6 +372,6 @@ InstitutionActions.propTypes = {
   onChange: PropTypes.func.isRequired
 };
 
-const mapContextToProps = ({ user }) => ({ user });
+const mapContextToProps = ({ addError, addSuccess, user }) => ({ user, addError, addSuccess });
 
 export default withContext(mapContextToProps)(injectIntl(injectSheet(styles)(InstitutionActions)));
